fix: update broken import paths after service folder reorganisation
- Fix referenceData.zod imports to use referenceData/referenceData.zod path
- Fix assignmentDefinition.zod imports to use assignmentDefinition/assignmentDefinition.zod path
- Fix backendConfiguration.zod imports to use backendConfiguration/backendConfiguration.zod path
