+++ src/frontend/src/test/assignmentDefinition/sharedTestFixtures.ts
@@ -4,8 +4,8 @@
  * Reduces duplication and ensures consistency across test suites.
  */
 
-import type { AssignmentDefinition } from '../../services/assignmentDefinition.zod.ts';
-import type { AssignmentTopic, YearGroup } from '../../services/referenceData.zod.ts';
+import type { AssignmentDefinition } from '../../services/assignmentDefinition/assignmentDefinition.zod';
+import type { AssignmentTopic, YearGroup } from '../../services/referenceData/referenceData.zod';
 
 /**
  * Standard mock topics for testing.
